Move compactor to separate file.
